refactor: ResumeCategoryDetails 파일 분리
