+++ src/components/organisms/ResumeCategoryDetails/index.ts
@@ -1,0 +1,1 @@
+export { default as ResumeCategoryDetails } from './ResumeCategoryDetails';
